+++ docs/changes.rst
@@ -16,6 +16,28 @@ Change log
 ==========
 
 
+Version 0.6.1.dev1
+------------------
+
+Released: not yet
+
+**Incompatible changes:**
+
+**Deprecations:**
+
+**Bug fixes:**
+
+**Enhancements:**
+
+**Cleanup:**
+
+**Known issues:**
+
+* See `list of open issues`_.
+
+.. _`list of open issues`: https://github.com/andy-maier/immutable-views/issues
+
+
 Version 0.6.0
 -------------